==> Chapter10/01_Neo4j_bindings.ipynb <==
# Graph Database Connection

In the following, we will show you how to connect and query data on Neo4j, using python. 

**IMPORTANT NOTE**

This notebook requires that you have access to a working version of Neo4j. In order to install Neo4j locally, we advise you to refer to the Neo4j webpage (https://neo4j.com/download/) or to use docker (https://hub.docker.com/_/neo4j).

with open("./movieCreationQuery.txt", "rb") as fid:
    lines = fid.readlines()

query = " ".join([line.decode("utf-8").replace("\n", "") for line in lines])

from neo4j import GraphDatabase

import os
host = os.environ.get("NEO4J_HOST", "localhost")

uri = f"neo4j://{host}:7687"
driver = GraphDatabase.driver(uri, auth=("neo4j", "neo5j"))

def run_query(tx, query):
    return list(tx.run(query))

with driver.session() as session:
    session.write_transaction(run_query, query)

Query

query = "MATCH (n) RETURN count(*)"

with driver.session() as session:
    result = session.read_transaction(run_query, query)
[r for r in result]

### Using `graphdatascience`

import graphdatascience

from graphdatascience import GraphDataScience

import os
host = os.environ.get("NEO4J_HOST", "localhost")

uri = f"bolt://{host}:7687"
gds = GraphDataScience(uri, auth=("neo4j", "neo5j"))

gds.run_cypher("MATCH (n) RETURN count(*);") 

### Using the analytics capabilities of `graphdatascience`

G = gds.graph.load_cora()

gds.graph.list()

G=gds.graph.get("cora")

G.node_properties()

gds.graph.nodeProperty.stream(gds.graph.get("cora"), node_property="subject")

pr_result = gds.pageRank.mutate(G, mutateProperty="pagerank")

print(f"Compute millis: {pr_result['computeMillis']}")
print(f"Node properties written: {pr_result['nodePropertiesWritten']}")
print(f"Centrality distribution: {pr_result['centralityDistribution']}")

G.node_properties()

gds.graph.nodeProperties.stream(G, ["pagerank"], separate_property_columns=True)
# gds.graph.nodeProperties.write(G, ["pagerank"])

### Delete datasets

gds.graph.drop("cora")

with driver.session() as session:
    result = session.write_transaction(run_query, "MATCH (n)-[e]-() DELETE n, e")
