==> Chapter10/02_JanusGraph_Gremlin.ipynb <==
# JanusGraph and Gremlin queries

from gremlin_python import statics
from gremlin_python.structure.graph import Graph
from gremlin_python.process.graph_traversal import __
from gremlin_python.driver.driver_remote_connection import DriverRemoteConnection
from gremlin_python.driver.serializer import GraphSONSerializersV3d0

import nest_asyncio
nest_asyncio.apply()

from gremlin_python.process.anonymous_traversal import traversal

import os
host = os.environ.get("JANUSGRAPH_HOST", "localhost")

connection = DriverRemoteConnection(f"ws://{host}:8182/gremlin", "g", message_serializer=GraphSONSerializersV3d0())

graph = Graph()
g = graph.traversal().withRemote(connection)

g.addV('student').property('name', 'Jeffery').property('GPA', 4.0).next()

g.addV('student').property('name', 'Robert').property('GPA', 3.0).next()

v1, v2 = g.V().to_list()

v1

v2

g.addE("FRIEND_OF").from_(v1).to(v2).property("since", "2014").next()

g\
  .addV('student').property('name', 'Claire').property('GPA', 3.9).as_("n1")\
  .addV('student').property('name', 'Lisa').property('GPA', 3.6).as_("n2")\
  .addE("FRIEND_OF").from_("n1").to("n2").property("since", "2014")\
  .iterate()

g.V().to_list()

g.E().to_list()

g.V().drop().to_list()

g.E().drop().to_list()

### Import Karate Club Graph 

import networkx

import pandas as pd

nodes = networkx.karate_club_graph().nodes
nodes = pd.DataFrame.from_records([{"id": node} | nodes[node] for node in nodes]).set_index("id")

edges = networkx.karate_club_graph().edges
edges = pd.DataFrame.from_records([{"id": edge} | edges[edge] for edge in edges]).set_index("id")

nodes.head()

edges.head()

Graph Generation 

from gremlin_python.process.graph_traversal import GraphTraversalSource

from functools import reduce

def build_node_query(agg: GraphTraversalSource, id: str, label: str, properties:dict):
    id_str = str(id)
    agg = agg.add_v(label).property("id", id_str)
    for k, v in properties.items():
        agg.property(k, v)
    return agg.as_(f"n_{id_str}")

def build_edge_query(agg: GraphTraversalSource, id: tuple[str,str], label: str, properties:dict):
    source_str = str(id[0])
    target_str = str(id[1])
    edge = agg\
        .V().has("id", str(source_str)).as_("source")\
        .V().has("id", str(target_str)).as_("target")\
        .addE(label).from_("source").to("target")
    for k, v in properties.items():
        edge.property(k, v)
    return edge.as_(f"edge_{source_str}_{target_str}")

_ = reduce(lambda g, node: build_node_query(g, node[0], "Person", node[1].to_dict()), nodes.iterrows(), g).iterate()

_ = reduce(lambda g, edge: build_edge_query(g, edge[0], "FRIEND_OF", edge[1].to_dict()), edges.iterrows(), g).iterate()

g.V().has("club", "Mr. Hi").out("FRIEND_OF").has("club", 'Officer').count().next()

g.V().has("club", "Mr. Hi").out("FRIEND_OF").has("club", 'Officer').values("id").to_list()

g.V().has("club", "Officer").in_("FRIEND_OF").has("club", 'Mr. Hi').values("id").to_list()

### Drop databases

g.V().drop().to_list()

g.E().drop().to_list()

### Import Movie Graph

nodes = pd.read_pickle("nodes.pkl")

edges = pd.read_pickle("edges.pkl")

nodes.head()

edges.head()

Creation of edges and nodes batch by batch

from itertools import islice

def batched(iterable, n):
    "Batch data into lists of length n. The last batch may be shorter."
    # batched('ABCDEFG', 3) --> ABC DEF G
    it = iter(iterable)
    while True:
        batch = list(islice(it, n))
        if not batch:
            return
        yield batch

def create_from_batch(builder, iterable, batch_size):
    for batch in batched(iterable, batch_size):
        _ = reduce(lambda g, item: builder(g, item[0], item[1]["label"], item[1]["props"]), batch, g).iterate()

create_from_batch(build_node_query, nodes.iterrows(), 10)

create_from_batch(build_edge_query, edges.iterrows(), 10)

g.V().count().next()

g.E().count().next()

g.V().has('Person', 'name', 'Keanu Reeves').out("ACTED_IN").in_("ACTED_IN").values("name").dedup().to_list()

g.V().drop().to_list()

g.E().drop().to_list()
